refactor: rename task_tests to back p2p_task_test
